tcpresponder/requirements.txt
fastapi
uvicorn
aiohttp
pydantic
requests
pandas
python-dotenv